Refactoring Record.serialize methods
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@718810 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFCellStyle.java
@@ -1015,9 +1015,7 @@ public class HSSFCellStyle implements CellStyle
     	if(workbook != source.workbook) {
 			// Then we need to clone the format string,
 			//  and update the format record for this
-    		short fmt = workbook.createFormat(
-    				source.getDataFormatString()
-    		);
+    		short fmt = (short)workbook.createFormat(source.getDataFormatString() );
     		setDataFormat(fmt);
 			
 			// Finally we need to clone the font,
